+++ frontend/src/App.vue
@@ -1,15 +1,17 @@
 <template>
   <div id="app">
     <!--<theme-button style="display:none" />-->
+    <refresh></refresh>
     <router-view/>
   </div>
 </template>
 
 <script>
 import ThemeButton from '@/components/ThemeButton.vue'
+import Refresh from './components/Refresh.vue'
 export default {
   name: 'App',
-  components: { ThemeButton }
+  components: { ThemeButton, Refresh }
 }
 </script>