+++ frontend/src/components/inbox/NumberList.vue
@@ -47,6 +47,16 @@
             </b-dropdown-item-button>
           </b-dropdown>
         </div>
+        <!--<div class="icons mt-2">
+          <b-dropdown size="sm" class="dropDown"  variant="primary" toggle-class="text-decoration-none" no-caret>
+          <template #button-content>
+            <b-icon icon="three-dots-vertical" font-scale="1"></b-icon>
+          </template>
+          <b-dropdown-item href="#">Action</b-dropdown-item>
+          <b-dropdown-item href="#">Another action</b-dropdown-item>
+          <b-dropdown-item href="#">Something else here...</b-dropdown-item>
+        </b-dropdown>
+        </div>-->
       </div>
     </div>
     <div class="wrap-search">
@@ -94,7 +104,8 @@
         ></b-icon>
         <div class="contact-preview">
           <div class="contact-text">
-            <h1 class="font-name">{{ item._id }}</h1>
+            <h1 class="font-name" v-if="item.contact">{{item.contact.first_name}} {{item.contact.last_name}}</h1>
+            <h1 v-else class="font-name">{{ item._id }}</h1>
             <p class="font-preview">{{ getValidString(item.message)  }}</p>
           </div>
         </div>
@@ -331,7 +342,7 @@ import ThemeButton from '@/components/ThemeButton.vue'
 import ProfileView from '@/components/setting/ProfileView.vue'
 import Contact from '@/components/setting/Contact.vue'
 import { required } from 'vuelidate/lib/validators'
-import { get } from '../../core/module/common.module'
+import { get, post } from '../../core/module/common.module'
 export default {
   components: {
     ProfileView, ThemeButton, Contact
@@ -418,21 +429,17 @@ export default {
       return newStr2
     },
     getOneProfile () {
-      // eslint-disable-next-line no-undef
-      axios.post(`${this.baseurl}/profile/getdata-one`, { setting: this.activeProfile._id }, this.headers)
-        .then(response => {
-          this.activeProfile = response.data.data
+      var request = {
+        data: { setting: this.activeProfile._id },
+        url: 'profile/getdata-one'
+      }
+      this.$store
+        .dispatch(post, request)
+        .then((response) => {
+          this.activeProfile = response.data
         })
-        .catch(error => {
-          if (error.response.status === 401) {
-            this.$swal({
-              icon: 'error',
-              title: 'Oops...',
-              text: error.response.data.message
-            })
-            this.$cookie.delete('access_token')
-            this.$router.push('/')
-          }
+        .catch((e) => {
+          console.log(e)
         })
     },
     refreshProfile () {
@@ -468,49 +475,40 @@ export default {
     },
     getNumberList () {
       this.numbers = []
-      // eslint-disable-next-line no-undef
-      axios.post(`${this.baseurl}/setting/sms-number-list`, {user: this.userdata._id, setting: this.activeProfile._id}, this.headers)
-        .then(response => {
-          this.numbers = response.data
+      var request = {
+        data: {user: this.userdata._id, setting: this.activeProfile._id},
+        url: 'setting/sms-number-list'
+      }
+      this.$store
+        .dispatch(post, request)
+        .then((response) => {
+          this.numbers = response
           this.messageListLoader = false
         })
-        .catch(error => {
-          if (error.response.status === 401) {
-            this.$swal({
-              icon: 'error',
-              title: 'Oops...',
-              text: error.response.data.message
-            })
-            this.$cookie.delete('access_token')
-            this.$router.push('/')
-          }
+        .catch((e) => {
+          console.log(e)
         })
     },
     getSetting () {
-      // eslint-disable-next-line no-undef
-      axios.post(`${this.baseurl}/setting/get-setting`, {user: this.userdata._id, setting: this.activeProfile._id}, this.headers)
-        .then(response => {
-          // this.numbers = response.data
-          if (response.data.data) {
-            this.user = response.data.data
-            this.user.twilio_number = response.data.data.number
-            if (response.data.data.number) {
+      var request = {
+        data: {user: this.userdata._id, setting: this.activeProfile._id},
+        url: 'setting/get-setting'
+      }
+      this.$store
+        .dispatch(post, request)
+        .then((response) => {
+          if (response.data) {
+            this.user = response.data
+            this.user.twilio_number = response.data.number
+            if (response.data.number) {
               // this.socket.emit('join_channel', this.user.number)
             }
-            this.getNumbers(response.data.data.type)
-            this.selected = response.data.data.type
+            this.getNumbers(response.data.type)
+            this.selected = response.data.type
           }
         })
-        .catch(error => {
-          if (error.response.status === 401) {
-            this.$swal({
-              icon: 'error',
-              title: 'Oops...',
-              text: error.response.data.message
-            })
-            this.$cookie.delete('access_token')
-            this.$router.push('/')
-          }
+        .catch((e) => {
+          console.log(e)
         })
     },
     deleteProfile () {
@@ -523,40 +521,38 @@ export default {
         denyButtonText: `No`
       }).then((result) => {
         if (result.isConfirmed) {
-          // eslint-disable-next-line no-undef
-          axios.post(`${this.baseurl}/profile/delete-profile`, {user: this.userdata._id, profile_id: this.activeProfile._id}, this.headers)
-            .then(response => {
-              this.$swal({
-                icon: 'success',
-                title: 'Success',
-                text: 'Profile deleted successfully!'
-              })
-              this.user.api_key = ''
-              this.user.number = ''
-              this.user.twilio_sid = ''
-              this.user.twilio_token = ''
-              this.user.twilio_number = ''
-              this.tNumbers = []
-              this.twilioNumbers = []
-              this.activeProfile = response.data.data
-              localStorage.removeItem('activeProfile')
-              this.$refs['my-modal'].hide()
-              this.$refs.childComponent.getallProfile()
-              var $this = this
-              setTimeout(function () {
-                $this.$refs.childComponent.activeFirstProfile()
-              }, 2000)
-            })
-            .catch(error => {
-              if (error.response.status === 401) {
+          var request = {
+            data: {user: this.userdata._id, profile_id: this.activeProfile._id},
+            url: 'profile/delete-profile'
+          }
+          this.$store
+            .dispatch(post, request)
+            .then((response) => {
+              if (response.data) {
                 this.$swal({
-                  icon: 'error',
-                  title: 'Oops...',
-                  text: error.response.data.message
+                  icon: 'success',
+                  title: 'Success',
+                  text: 'Profile deleted successfully!'
                 })
-                this.$cookie.delete('access_token')
-                this.$router.push('/')
+                this.user.api_key = ''
+                this.user.number = ''
+                this.user.twilio_sid = ''
+                this.user.twilio_token = ''
+                this.user.twilio_number = ''
+                this.tNumbers = []
+                this.twilioNumbers = []
+                this.activeProfile = response.data
+                localStorage.removeItem('activeProfile')
+                this.$refs['my-modal'].hide()
+                this.$refs.childComponent.getallProfile()
+                var $this = this
+                setTimeout(function () {
+                  $this.$refs.childComponent.activeFirstProfile()
+                }, 2000)
               }
+            })
+            .catch((e) => {
+              console.log(e)
             })
         } else if (result.isDenied) {
           // eslint-disable-next-line no-undef
@@ -574,9 +570,13 @@ export default {
         denyButtonText: `No`
       }).then((result) => {
         if (result.isConfirmed) {
-          // eslint-disable-next-line no-undef
-          axios.post(`${this.baseurl}/setting/delete-key`, {user: this.userdata._id, profile_id: this.activeProfile._id}, this.headers)
-            .then(response => {
+          var request = {
+            data: {user: this.userdata._id, profile_id: this.activeProfile._id},
+            url: 'setting/delete-key'
+          }
+          this.$store
+            .dispatch(post, request)
+            .then((response) => {
               this.$swal({
                 icon: 'success',
                 title: 'Success',
@@ -589,22 +589,12 @@ export default {
               this.user.twilio_number = ''
               this.tNumbers = []
               this.twilioNumbers = []
-              this.activeProfile = response.data.data
+              this.activeProfile = response.data
               this.$refs.childComponent.getallProfile()
             })
-            .catch(error => {
-              if (error.response.status === 401) {
-                this.$swal({
-                  icon: 'error',
-                  title: 'Oops...',
-                  text: error.response.data.message
-                })
-                this.$cookie.delete('access_token')
-                this.$router.push('/')
-              }
+            .catch((e) => {
+              console.log(e)
             })
-          // eslint-disable-next-line no-undef
-          // this.$swal.fire('Deleted!', '', 'success')
         } else if (result.isDenied) {
           // eslint-disable-next-line no-undef
           this.$swal.fire('setting not deleted', '', 'info')
@@ -614,25 +604,26 @@ export default {
     getNumbers (type) {
       var settings = this.user
       settings.type = type
-      // eslint-disable-next-line no-undef
-      axios.post(`${this.baseurl}/setting/get-number`, settings, this.headers)
-        .then(response => {
+      var request = {
+        data: settings,
+        url: 'setting/get-number'
+      }
+      if (type === 'telnyx') {
+        this.tNumbers = []
+      } else {
+        this.twilioNumbers = []
+      }
+      this.$store
+        .dispatch(post, request)
+        .then((response) => {
           if (type === 'telnyx') {
-            this.tNumbers = response.data.data.data
+            this.tNumbers = response.data.data
           } else {
-            this.twilioNumbers = response.data.data
+            this.twilioNumbers = response.data
           }
         })
-        .catch(error => {
-          if (error.response.status === 401) {
-            this.$swal({
-              icon: 'error',
-              title: 'Oops...',
-              text: error.response.data.message
-            })
-            this.$cookie.delete('access_token')
-            this.$router.push('/')
-          }
+        .catch((e) => {
+          console.log(e)
         })
     },
     handleSubmit (e) {
@@ -670,36 +661,25 @@ export default {
           setting: this.activeProfile._id,
           profile: this.user.profile
         }
-        // eslint-disable-next-line no-undef
-        axios.post(`${this.baseurl}/setting/create`, sendData, this.headers)
-          .then(response => {
+        var request = {
+          data: sendData,
+          url: 'setting/create'
+        }
+        this.$store
+          .dispatch(post, request)
+          .then((response) => {
             this.$swal({
               icon: 'success',
               title: 'Success',
               text: 'Settings saved successfully!'
             })
             this.$refs['my-modal'].hide()
-            this.activeProfile = response.data.data
+            this.activeProfile = response.data
             this.$refs.childComponent.getallProfile()
             this.$v.$reset()
           })
-          .catch(error => {
-            if (error.response.status === 401) {
-              this.$swal({
-                icon: 'error',
-                title: 'Oops...',
-                text: error.response.data.message
-              })
-              this.$cookie.delete('access_token')
-              this.$router.push('/')
-            }
-            if (error.response.status === 400) {
-              this.$swal({
-                icon: 'error',
-                title: 'Oops...',
-                text: error.response.data.message
-              })
-            }
+          .catch((e) => {
+            console.log(e)
           })
       }
     }